Added new logic to check for incomplete AGR records.
Added new logic to check for incomplete AGR records. For example, when a Gage is offline for several days (0 count) records are written to the database. After the Gage is powered on an existing record is flagged as complete (at the end of a shift) where the start counts are all 0s because and the end counts are all real numbers from the PLC, then, when the math is applied the Total, AGR and Net values are extremely high and are completely out of range.